+++ java/src/main/java/gherkin/formatter/JSONPrettyFormatter.java
@@ -1,9 +1,6 @@
 package gherkin.formatter;
 
 
-import gherkin.deps.com.google.gson.Gson;
-import gherkin.deps.com.google.gson.GsonBuilder;
-
 @Deprecated
 public class JSONPrettyFormatter extends JSONFormatter {
     public JSONPrettyFormatter(Appendable out) {